fix RecordsProducer vs RecordsHandler stream env var handling
Setting an env var for SK_STREAM causes tests to fail. This is because
RecordsProducer set stream_name to the env var if nothing was passed in.
This is different behaviour to Ingester PSC, which sets it only in
RecordsHandler.

Fix by using Ingester PSC approach.

Note that this is another example of subtle drift between repos caused
by code duplication.