[#217] update env vars for dev Kinesis
IAM permissions have been granted for new dev Kinesis streams. This
updates env vars so everything works out of the box.

It was necessary to set DK_CVR_USERNAME, DK_CVR_PASSWORD.

It did not appear necessary to set BODS_S3_BUCKET_NAME, similar to
Ingester SK. I'm not sure why it was set or why there might be a
difference, so I'm unsetting it until it's clear.

Unlike for Ingester SK, it appears necessary to set BODS_S3_BUCKET_NAME.
However, that's not for all records (perhaps only because of large
data for this datasource?). So, I'm leaving it unset, since the basic
operation works in dev (11200 records) before crashing. Setting would
also require a review of the dev permissions, if that bucket isn't
read-only. So I'm leaving it out until it's needed.